requirements.txt
appdirs==1.4.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bleach==1.5.0
certifi==2017.1.23
configparser==3.5.0
cycler==0.10.0
Cython==0.25.2
dask==0.13.0
decorator==4.0.11
entrypoints==0.2.2
enum34==1.1.6
flickr-api==0.5
functools32==3.2.3.post2
h5py==2.6.0
html5lib==0.9999999
ipykernel==4.5.2
ipython==5.2.2
ipython-genutils==0.1.0
ipywidgets==5.2.2
Jinja2==2.9.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.1.0
jupyter-core==4.2.1
leveldb==0.194
MarkupSafe==0.23
matplotlib==2.0.0
mistune==0.7.3
nbconvert==5.1.1
nbformat==4.2.0
networkx==1.11
nose==1.3.7
notebook==4.4.1
numpy==1.12.0
oauth==1.0.1
olefile==0.44
packaging==16.8
pandas==0.19.2
pandocfilters==1.4.1
pathlib2==2.2.1
pbr==1.10.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.13
protobuf==3.2.0
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.1.10
python-dateutil==1.5
python-gflags==3.1.1
pytz==2016.10
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.2.1
requests==2.13.0
scandir==1.4
scikit-image==0.12.3
scipy==0.18.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
stevedore==1.20.0
subprocess32==3.2.7
terminado==0.6
testpath==0.3
toolz==0.8.2
tornado==4.4.2
traitlets==4.3.1
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
wcwidth==0.1.7
widgetsnbextension==1.2.6
